Most menu items now perform some action (although those actions aren't necessarily complete yet).
Selecting the bankswitch type now actually changes it for the currently
selected ROM.  Still TODO is issue a warning when overriding the detected
type, and perhaps allowing the new type to be saved indefinitely.

There is now a RED icon indicating that the KrokCart wasn't found,
similar to what's in the Windows KrokCommander software.