Added download verification.
Save serial port where KrokCart is detected for the next time it should
be opened.  This location is then checked first, potentially eliminating
a search through all ports.  If the device isn't found at the saved
location, as full search is performed as before.

Added better error handling.  If download or verify fails, the issue is
reported with more informative messages.